spring-boot-aviator/src/test/java/com/eric/service/SomeBusinessServiceTest.java
package com.eric.service;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;
import static org.mockito.Mockito.*;
import static org.junit.Assert.*;

import java.util.HashMap;
import java.util.Map;

@RunWith(MockitoJUnitRunner.class)
public class SomeBusinessServiceTest {

    @Mock
    private ExpressionService expressionService;

    @InjectMocks
    private SomeBusinessService someBusinessService;

    @Before
    public void setUp() {
        // This setup is now optional as MockitoJUnitRunner and @Mock, @InjectMocks take care of initialization.
    }

    @Test
    public void testSomeBusinessMethod() {
        // Setup
        Map<String, Object> variables = new HashMap<>();
        variables.put("x", 10);
        variables.put("y", 20);
        when(expressionService.evaluate(anyString(), eq(variables))).thenReturn(30);

        // Exercise
        someBusinessService.someBusinessMethod();

        // Verify
        verify(expressionService).evaluate(eq("x + y"), eq(variables));

        // Note: Since the method someBusinessMethod does not return a value and only prints the result,
        // we cannot assert the output directly. Instead, we verify that the expressionService.evaluate
        // method was called with the expected parameters. In a real-world scenario, you might want to
        // refactor the code to make the result of the calculation available for assertion.
    }

    @Test
    public void testEvaluateWithDifferentVariables() {
        // Setup for a different scenario
        Map<String, Object> differentVariables = new HashMap<>();
        differentVariables.put("a", 5);
        differentVariables.put("b", 15);
        when(expressionService.evaluate(anyString(), eq(differentVariables))).thenReturn(20);

        // Exercise
        someBusinessService.someBusinessMethod(); // Assuming method can handle different variables

        // Verify
        verify(expressionService).evaluate(eq("x + y"), eq(differentVariables));

        // This test highlights the ability of your service to handle different variable sets
        // which is a good practice for writing unit tests - to test the edge cases and different scenarios as well.
    }
}
